# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Vidalytics Multi-Agent API",
    "version": "2.0.0",
...
    "summary": "AI-powered YouTube optimization platform",
    "contact": {
      "name": "Vidalytics Support",
      "url": "https://github.com/Vidalytics/api",
      "email": "support@Vidalytics.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8888",
      "description": "Development server"
    },
    {
      "url": "https://api.Vidalytics.com",
      "description": "Production server"
    }
  ],
  "security": [
    {
      "SessionCookie": []
    },
    {
      "BearerAuth": []
    }
  ],
...
  "components": {
    "securitySchemes": {
      "SessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "Vidalytics_session",
        "description": "Session-based authentication using secure HTTP-only cookies"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Bearer token authentication using session ID"
      }
    },
    "schemas": {
      "StandardResponse": {
        "type": "object",
        "required": [
          "success",
          "message",
          "timestamp"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Indicates if the request was successful"
          },
          "message": {
            "type": "string",
            "description": "Human-readable message describing the result"
          },
          "data": {
            "type": "object",
            "description": "Response data (structure varies by endpoint)"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDetails"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "ISO 8601 timestamp of the response"
          }
        }
      },
      "ErrorDetails": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "Error code for programmatic handling"
          },
          "details": {
            "type": "string",
            "description": "Detailed error description"
          },
          "field": {
            "type": "string",
            "description": "Field name for validation errors"
          }
        }
      },
...
    },
    "responses": {
      "UnauthorizedError": {
        "description": "Authentication credentials are missing or invalid",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/StandardResponse"
            },
            "example": {
              "success": false,
              "message": "Authentication required",
              "error": {
                "code": "AUTHENTICATION_REQUIRED",
                "details": "Valid session required to access this endpoint"
              },
              "timestamp": "2024-01-01T12:00:00Z"
            }
          }
        }
      },
...
      "ValidationError": {
        "description": "Request validation failed",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/StandardResponse"
            },
            "example": {
              "success": false,
              "message": "Validation error",
              "error": {
                "code": "VALIDATION_ERROR",
                "details": "Required field is missing",
                "field": "user_id"
              },
              "timestamp": "2024-01-01T12:00:00Z"
            }
          }
        }
      },
...
      "ServerError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/StandardResponse"
            },
            "example": {
              "success": false,
              "message": "Internal server error",
              "error": {
                "code": "INTERNAL_SERVER_ERROR",
                "details": "An unexpected error occurred"
              },
              "timestamp": "2024-01-01T12:00:00Z"
            }
          }
        }
      }
    },
    "parameters": {
...
      "sort_by": {
        "name": "sort_by",
        "in": "query",
        "description": "Field to sort by",
        "required": false,
        "schema": {
          "type": "string",
          "default": "created_at"
        }
      },
      "order": {
        "name": "order",
        "in": "query",
        "description": "Sort order",
        "required": false,
        "schema": {
          "type": "string",
          "enum": [
            "asc",
            "desc"
          ],
          "default": "desc"
        }
      }
    },
    "examples": {
...
      "PillarList": {
        "summary": "List of content pillars",
        "value": {
          "success": true,
          "message": "Pillars retrieved successfully",
          "data": {
            "pillars": [
              {
                "id": "pillar_123",
                "name": "Educational Content",
                "description": "Tutorials, how-to guides, and educational videos",
                "color": "#3B82F6",
                "created_at": "2024-01-01T10:00:00Z",
                "video_count": 25,
                "analytics": {
                  "total_views": 150000,
                  "average_ctr": 0.052,
                  "average_retention": 0.68,
                  "engagement_rate": 0.071
                }
              }
            ],
            "total_count": 5
          },
          "timestamp": "2024-01-01T12:00:00Z"
        }
      }
    }
  },
  "paths": {
...
    "/api/pillars": {
      "get": {
        "tags": [
          "pillars"
        ],
        "summary": "List content pillars",
        "description": "Get all content pillars for the authenticated user",
        "operationId": "listPillars",
        "security": [
          {
            "SessionCookie": []
          },
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "include_videos",
            "in": "query",
            "description": "Include assigned videos in response",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            }
          },
          {
            "name": "include_analytics",
            "in": "query",
            "description": "Include pillar performance analytics",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            }
          },
          {
            "$ref": "#/components/parameters/sort_by"
          },
          {
            "$ref": "#/components/parameters/order"
          }
        ],
        "responses": {
          "200": {
            "description": "Pillars retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse"
                },
                "examples": {
                  "pillar_list": {
                    "$ref": "#/components/examples/PillarList"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "post": {
        "tags": [
          "pillars"
        ],
        "summary": "Create new content pillar",
        "description": "Create a new content pillar for organizing videos",
        "operationId": "createPillar",
        "security": [
          {
            "SessionCookie": []
          },
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Pillar name"
                  },
                  "description": {
                    "type": "string",
                    "description": "Pillar description"
                  },
                  "color": {
                    "type": "string",
                    "pattern": "^#[0-9A-Fa-f]{6}$",
                    "description": "Pillar color (hex format)"
                  },
                  "metadata": {
                    "type": "object",
                    "description": "Custom metadata"
                  }
                }
              },
              "examples": {
                "create_pillar": {
                  "summary": "Create educational pillar",
                  "value": {
                    "name": "Educational Content",
                    "description": "Tutorials and how-to videos",
                    "color": "#3B82F6",
                    "metadata": {
                      "target_audience": "beginners",
                      "content_frequency": "weekly"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Pillar created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
  }
}